Add new BACDL_BSC datalink as a BACnet/SC Node
This commit adds a new BACDL_BSC datalink, at the same level of
abstraction as other BACnet datalinks.  It makes use of libwebsockets
for the WebSocket implementation, which is also event based and seems
to integrate well into the overall BACnet stack framework.

To use, configre the stack using
 `cmake . -DBAC_ROUTING=0 -DBACDL_BIP=ON -DBACDL_BSC=ON`

And then configure your environment.
 BACNET_IFACE=wss://localhost:4443
 BACNET_SC_CERTFILE=<certificate path>
 BACNET_SC_KEYFILE=<certificate path>
 BACNET_SC_CAFILE=<certificate path>

In this initial PoC push only a basic SC Node connection is
implemented, and there are many TODOS:

 * VMAC and device UUID are hard coded
 * Failover to a secondary hub as specified in the standard are not implemented.
 * No thought given to how this will work in different ports other than Linux (does libwebsockets work)
 * The only BVLL Response message parsed is the NACK you get when connecting to a hub sometimes
 * Peer-to-peer connections are not implemented

It has been lightly tested with the BACnet/SC reference implementation
(https://sourceforge.net/projects/bacnet-sc-reference-stack/), as well
as Delta entelliWeb.